# api/openapi-spec.v2.json
{
"openapi": "3.0.1",
"info": {
"title": "Competition Engine",
"version": "2.0"
},
"servers": [
{
"url": "https://api-live.euroleague.net"
}
],
"paths": {
...
"/v2/clubs": {
"get": {
"tags": [
"Clubs"
],
"summary": "Get all registered clubs",
"parameters": [
{
"name": "Limit",
"in": "query",
"schema": {
"maximum": 2147483647,
"minimum": 0,
"type": "integer",
"format": "int32"
}
},
{
"name": "Offset",
"in": "query",
"schema": {
"maximum": 2147483647,
"minimum": 0,
"type": "integer",
"format": "int32"
}
},
{
"name": "hasParentClub",
"in": "query",
"description": "",
"schema": {
"type": "boolean"
}
},
{
"name": "search",
"in": "query",
"description": "",
"schema": {
"type": "string"
}
}
],
"responses": {
"200": {
"description": "Success",
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/ClubModelPagingResultModel"
}
},
"application/xml": {
"schema": {
"$ref": "#/components/schemas/ClubModelPagingResultModel"
}
}
}
},
"401": {
"description": "Unauthorized"
},
"403": {
"description": "Forbidden"
}
},
"security": [
{
"OAuth2": [
"euroleagueapi"
]
}
]
}
},
"/v2/clubs/{clubCode}": {
"get": {
"tags": [
"Clubs"
],
"summary": "Get a club by its code",
"parameters": [
{
"name": "clubCode",
"in": "path",
"description": "",
"required": true,
"schema": {
"type": "string"
}
}
],
"responses": {
"200": {
"description": "Success",
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/ClubModel"
}
},
"application/xml": {
"schema": {
"$ref": "#/components/schemas/ClubModel"
}
}
}
},
"401": {
"description": "Unauthorized"
},
"403": {
"description": "Forbidden"
}
},
"security": [
{
"OAuth2": [
"euroleagueapi"
]
}
]
}
},
...
"/v2/clubs/{clubCode}/info": {
"get": {
"tags": [
"Clubs"
],
"summary": "Get a club by its code",
"parameters": [
{
"name": "clubCode",
"in": "path",
"description": "",
"required": true,
"schema": {
"type": "string"
}
}
],
"responses": {
"200": {
"description": "Success",
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/ClubInfoModel"
}
},
"application/xml": {
"schema": {
"$ref": "#/components/schemas/ClubInfoModel"
}
}
}
},
"401": {
"description": "Unauthorized"
},
"403": {
"description": "Forbidden"
}
},
"security": [
{
"OAuth2": [
"euroleagueapi"
]
}
]
}
},
...
"/v2/competitions/{competitionCode}/seasons/{seasonCode}/clubs": {
"get": {
"tags": [
"SeasonClubs"
],
"summary": "Get a competition season clubs",
"parameters": [
{
"name": "competitionCode",
"in": "path",
"description": "",
"required": true,
"schema": {
"type": "string"
}
},
{
"name": "seasonCode",
"in": "path",
"description": "",
"required": true,
"schema": {
"type": "string"
}
},
{
"name": "Limit",
"in": "query",
"schema": {
"maximum": 2147483647,
"minimum": 0,
"type": "integer",
"format": "int32"
}
},
{
"name": "Offset",
"in": "query",
"schema": {
"maximum": 2147483647,
"minimum": 0,
"type": "integer",
"format": "int32"
}
},
{
"name": "search",
"in": "query",
"description": "",
"schema": {
"type": "string"
}
}
],
"responses": {
"200": {
"description": "Success",
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/SeasonClubDetailsModelPagingResultModel"
}
},
"application/xml": {
"schema": {
"$ref": "#/components/schemas/SeasonClubDetailsModelPagingResultModel"
}
}
}
},
"401": {
"description": "Unauthorized"
},
"403": {
"description": "Forbidden"
}
},
"security": [
{
"OAuth2": [
"euroleagueapi"
]
}
]
}
},
"/v2/competitions/{competitionCode}/seasons/{seasonCode}/clubs/{clubCode}": {
"get": {
"tags": [
"SeasonClubs"
],
"summary": "Get a competition season club",
"parameters": [
{
"name": "competitionCode",
"in": "path",
"description": "",
"required": true,
"schema": {
"type": "string"
}
},
{
"name": "seasonCode",
"in": "path",
"description": "",
"required": true,
"schema": {
"type": "string"
}
},
{
"name": "clubCode",
"in": "path",
"description": "",
"required": true,
"schema": {
"type": "string"
}
}
],
"responses": {
"200": {
"description": "Success",
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/SeasonClubDetailsModel"
}
},
"application/xml": {
"schema": {
"$ref": "#/components/schemas/SeasonClubDetailsModel"
}
}
}
},
"401": {
"description": "Unauthorized"
},
"403": {
"description": "Forbidden"
}
},
"security": [
{
"OAuth2": [
"euroleagueapi"
]
}
]
}
},
...
"/v2/competitions/{competitionCode}/seasons/{seasonCode}": {
"get": {
"tags": [
"Seasons"
],
"summary": "Get a competition season",
"parameters": [
{
"name": "competitionCode",
"in": "path",
"description": "",
"required": true,
"schema": {
"type": "string"
}
},
{
"name": "seasonCode",
"in": "path",
"description": "",
"required": true,
"schema": {
"type": "string"
}
}
],
"responses": {
"200": {
"description": "Success",
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/SeasonDetailsModel"
}
},
"application/xml": {
"schema": {
"$ref": "#/components/schemas/SeasonDetailsModel"
}
}
}
},
"401": {
"description": "Unauthorized"
},
"403": {
"description": "Forbidden"
}
},
"security": [
{
"OAuth2": [
"euroleagueapi"
]
}
]
}
},
...
},
"components": {
"schemas": {
...
"ClubInfoModel": {
"type": "object",
"properties": {
"info": {
"type": "string",
"nullable": true
}
},
"additionalProperties": false
},
...
"ClubModel": {
"type": "object",
"properties": {
"code": {
"type": "string",
"nullable": true
},
"name": {
"type": "string",
"nullable": true
},
"alias": {
"type": "string",
"nullable": true
},
"isVirtual": {
"type": "boolean"
},
"country": {
"$ref": "#/components/schemas/CountryModel"
},
"address": {
"type": "string",
"nullable": true
},
"website": {
"type": "string",
"nullable": true
},
"ticketsUrl": {
"type": "string",
"nullable": true
},
"twitterAccount": {
"type": "string",
"nullable": true
},
"instagramAccount": {
"type": "string",
"nullable": true
},
"facebookAccount": {
"type": "string",
"nullable": true
},
"venue": {
"$ref": "#/components/schemas/VenueModel"
},
"venueBackup": {
"$ref": "#/components/schemas/VenueModel"
},
"nationalCompetitionCode": {
"type": "string",
"nullable": true
},
"city": {
"type": "string",
"nullable": true
},
"president": {
"type": "string",
"nullable": true
},
"phone": {
"type": "string",
"nullable": true
},
"fax": {
"type": "string",
"nullable": true
},
"images": {
"type": "object",
"additionalProperties": {
"type": "string"
},
"nullable": true
}
},
"additionalProperties": false
},
...
"ClubModelPagingResultModel": {
"type": "object",
"properties": {
"data": {
"type": "array",
"items": {
"$ref": "#/components/schemas/ClubModel"
},
"nullable": true
},
"total": {
"type": "integer",
"format": "int32"
}
},
"additionalProperties": false
},
...
"CountryModel": {
"type": "object",
"properties": {
"code": {
"type": "string",
"nullable": true
},
"name": {
"type": "string",
"nullable": true
}
},
"additionalProperties": false
},
...
"SeasonClubDetailsModel": {
"type": "object",
"properties": {
"code": {
"type": "string",
"nullable": true
},
"name": {
"type": "string",
"nullable": true
},
"abbreviatedName": {
"type": "string",
"nullable": true
},
"editorialName": {
"type": "string",
"nullable": true
},
"tvCode": {
"type": "string",
"nullable": true
},
"isVirtual": {
"type": "boolean"
},
"images": {
"type": "object",
"additionalProperties": {
"type": "string"
},
"nullable": true
},
"sponsor": {
"type": "string",
"nullable": true
},
"clubPermanentName": {
"type": "string",
"nullable": true
},
"clubPermanentAlias": {
"type": "string",
"nullable": true
},
"country": {
"$ref": "#/components/schemas/CountryModel"
},
"address": {
"type": "string",
"nullable": true
},
"website": {
"type": "string",
"nullable": true
},
"ticketsUrl": {
"type": "string",
"nullable": true
},
"twitterAccount": {
"type": "string",
"nullable": true
},
"venueCode": {
"type": "string",
"nullable": true
},
"city": {
"type": "string",
"nullable": true
},
"president": {
"type": "string",
"nullable": true
},
"phone": {
"type": "string",
"nullable": true
}
},
"additionalProperties": false
},
"SeasonClubDetailsModelPagingResultModel": {
"type": "object",
"properties": {
"data": {
"type": "array",
"items": {
"$ref": "#/components/schemas/SeasonClubDetailsModel"
},
"nullable": true
},
"total": {
"type": "integer",
"format": "int32"
}
},
"additionalProperties": false
},
...
"SeasonClubModel": {
"type": "object",
"properties": {
"code": {
"type": "string",
"nullable": true
},
"name": {
"type": "string",
"nullable": true
},
"abbreviatedName": {
"type": "string",
"nullable": true
},
"editorialName": {
"type": "string",
"nullable": true
},
"tvCode": {
"type": "string",
"nullable": true
},
"isVirtual": {
"type": "boolean"
},
"images": {
"type": "object",
"additionalProperties": {
"type": "string"
},
"nullable": true
}
},
"additionalProperties": false
},
...
"SeasonDetailsModel": {
"type": "object",
"properties": {
"name": {
"type": "string",
"nullable": true
},
"code": {
"type": "string",
"nullable": true
},
"alias": {
"type": "string",
"nullable": true
},
"competitionCode": {
"type": "string",
"nullable": true
},
"year": {
"type": "integer",
"format": "int32"
},
"startDate": {
"type": "string",
"format": "date-time",
"nullable": true
},
"activationDate": {
"type": "string",
"format": "date-time",
"nullable": true
},
"endDate": {
"type": "string",
"format": "date-time",
"nullable": true
},
"winner": {
"$ref": "#/components/schemas/SeasonClubModel"
}
},
"additionalProperties": false
},
...
"VenueModel": {
"type": "object",
"properties": {
"name": {
"type": "string",
"nullable": true
},
"code": {
"type": "string",
"nullable": true
},
"capacity": {
"type": "integer",
"format": "int32",
"nullable": true
},
"address": {
"type": "string",
"nullable": true
},
"images": {
"type": "object",
"additionalProperties": {
"type": "string"
},
"nullable": true
},
"active": {
"type": "boolean"
},
"notes": {
"type": "string",
"nullable": true
}
},
"additionalProperties": false
},
...
}
}
}